Add files via upload
A data scraping project on the Amazon website using Python involves automated web crawling to extract valuable information such as product details, pricing, reviews, and ratings. To execute this, libraries like BeautifulSoup and requests are employed. Initially, a list of target URLs is generated, and Python's requests library fetches the webpage content. BeautifulSoup parses the HTML, allowing the extraction of specific elements using CSS selectors. Data is then structured and stored in a format like CSV or a database. To enhance efficiency, concurrency techniques and user-agent rotation may be implemented. Ethical considerations and adherence to Amazon's terms of service are essential throughout the project.